csp-fusion/requirements.txt
python-dateutil
requests
mysql-connector-python
openpyxl